downloads search: Enable text search on id, manufacturer, and features
I have repeatedly been tripped up by wanting to enter a manufacturer
name in the text search area.  While it is possible to open the
filter panel and select manufacturers, it is more convenient and intuitive
to simply enter this in the single text search box.

By concatenating several fields from the dataset (name, id, manufacturer,
and features), and then searching that it becomes possible to text-search
e.g., for "alligator" (board feature), "diary" (manufacturer) or
"metro_m4" (board id)