Continue trying to fix RSS feed absolute URLs
Originally, both the RSS feed and the board list pages used the same
board_image.html include; I added `| absolute_url` to this file for
the RSS feed, which functions better when all links are absolute.

Using `| absolute_url` is preferable to just prepending
`https://circuitpython.org` because it can give a correct URL when
previewing locally with `bundle run jekyll serve`.

This crossed with a change by @makermelissa to switch the feeds
to use a separate include, presumably (though I didn't specifically ask
her about it) to avoid having the absoulute URLs where they were not
needed.

In this commit, I remove the use of `absolute_url` in the include that's
used to generate regular pages, while switching from prepending
`https://circuitpython.org` to using `| absolute_url` on the include
that is used just for the feed.